requirements.txt
django>=2.0.4,<2.1
django-debug-toolbar>=1.9.1,<1.10
gunicorn>=19.7.1,<19.8
psycopg2
psycopg2-binary
whitenoise>=3.3.1,<3.4
python-telegram-bot==10.0.2,<10.1
lxml==4.2.1,<4.3
mysqlclient>=1.3.12,<1.4
